Reorganize the code to add CLI
